Rename ~25 MCP tools per cli-restructure.md v2 §16b
Path-prefix-verb-suffix convention. Reorganize files:
- spending.py + budget_status → reports.py (reports_spending_*, reports_budget_status)
- categorize.py → transactions_categorize.py + categories.py + merchants.py
- accounts_balances → accounts_balance_list
- transactions_recurring → transactions_recurring_list
- categorize_* → transactions_categorize_* or categories_* or merchants_*
- import_csv_preview → import_preview
- budget_status → reports_budget_status
