+++ src/moneybin/services/budget_service.py
@@ -43,8 +43,8 @@ class BudgetSetResult:
             data=self.to_dict(),
             sensitivity="low",
             actions=[
-                "Use budget_status to see spending vs budget",
-                "Use spending_by_category for category breakdown",
+                "Use reports_budget_status to see spending vs budget",
+                "Use reports_spending_by_category for category breakdown",
             ],
         )
 
@@ -85,7 +85,7 @@ class BudgetStatusResult:
             period=self.month,
             actions=[
                 "Use budget_set to adjust a budget target",
-                "Use spending_by_category for detailed category breakdown",
+                "Use reports_spending_by_category for detailed category breakdown",
             ],
         )
