+++ src/moneybin/services/spending_service.py
@@ -55,8 +55,8 @@ class SpendingSummary:
             sensitivity="low",
             period=self.period_label,
             actions=[
-                "Use spending_by_category for category breakdown",
-                "Use spending_compare to compare periods",
+                "Use reports_spending_by_category for category breakdown",
+                "Use reports_spending_compare to compare periods",
             ],
         )
 
@@ -98,7 +98,7 @@ class CategoryBreakdown:
             sensitivity="low",
             period=self.period_label,
             actions=[
-                "Use spending_merchants for merchant-level breakdown",
+                "Use reports_spending_merchants for merchant-level breakdown",
                 "Use transactions_search to see individual transactions in a category",
             ],
         )
