+++ tests/moneybin/test_mcp/test_visibility.py
@@ -46,7 +46,7 @@ async def test_core_tools_visible_at_connect() -> None:
 
     async with Client(mcp) as client:
         names = {t.name for t in await client.list_tools()}
-        assert "spending_summary" in names
+        assert "reports_spending_summary" in names
         assert "accounts_list" in names
 
 
@@ -57,7 +57,7 @@ async def test_extended_tools_hidden_at_connect() -> None:
 
     async with Client(mcp) as client:
         names = {t.name for t in await client.list_tools()}
-        assert "categorize_bulk" not in names
+        assert "transactions_categorize_bulk_apply" not in names
         assert "budget_set" not in names
 
 
@@ -73,7 +73,7 @@ async def test_discover_reveals_namespace_tools() -> None:
     async with Client(mcp) as client:
         await client.call_tool("moneybin_discover", {"domain": "categorize"})
         names = {t.name for t in await client.list_tools()}
-        assert "categorize_bulk" in names
+        assert "transactions_categorize_bulk_apply" in names
 
 
 @pytest.mark.asyncio
@@ -103,15 +103,15 @@ async def test_per_session_discover_isolated() -> None:
 
     async with Client(mcp) as client_a, Client(mcp) as client_b:
         before_a = {t.name for t in await client_a.list_tools()}
-        assert "categorize_bulk" not in before_a
+        assert "transactions_categorize_bulk_apply" not in before_a
 
         await client_a.call_tool("moneybin_discover", {"domain": "categorize"})
 
         after_a = {t.name for t in await client_a.list_tools()}
         visible_b = {t.name for t in await client_b.list_tools()}
 
-        assert "categorize_bulk" in after_a
-        assert "categorize_bulk" not in visible_b, (
+        assert "transactions_categorize_bulk_apply" in after_a
+        assert "transactions_categorize_bulk_apply" not in visible_b, (
             "Client B's tool visibility leaked from Client A's discover call — "
             "session isolation is broken."
         )
@@ -132,7 +132,9 @@ async def test_visibility_or_match_semantics() -> None:
     async with Client(mcp) as client:
         await client.call_tool("moneybin_discover", {"domain": "categorize"})
         names = {t.name for t in await client.list_tools()}
-        assert "categorize_bulk" in names, "categorize tools should be enabled"
+        assert "transactions_categorize_bulk_apply" in names, (
+            "categorize tools should be enabled"
+        )
         assert "budget_set" not in names, (
             "budget tools must remain hidden when only categorize is enabled — "
             "Visibility transform OR-match contract is broken."
@@ -201,4 +203,4 @@ async def test_hidden_tool_is_uncallable_via_tools_call() -> None:
 
     async with Client(mcp) as client:
         with pytest.raises(ToolError):
-            await client.call_tool("categorize_bulk", {})
+            await client.call_tool("transactions_categorize_bulk_apply", {})
